pyproject.toml
[tool.poetry]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = ["Kirill8769 <tkkrozn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.4"
django = "^5.0.6"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
django-multiselectfield = "^0.1.12"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
celery = "^5.4.0"
redis = "^5.0.6"
django-celery-beat = "^2.6.0"
requests = "^2.32.3"
coverage = "^7.5.3"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.0.0"
isort = "^5.13.2"

[tool.isort]
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
